+++ BrotBoxEngineTest/Tests/DataStructures/ArrayTest.h
@@ -20,53 +20,53 @@ namespace bbe
 				assertUnequals(constArr.getRaw(), nullptr);
 				assertEquals(constArr.getLength(), 16);
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					arr[i] = i + 1000;
 				}
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					assertEquals(arr[i], i + 1000);
 				}
 
 				const bbe::Array<int, 16> constArr2 = arr;
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					assertEquals(arr[i], i + 1000);
 				}
-				for (int i = 0; i < constArr2.getLength(); i++)
+				for (size_t i = 0; i < constArr2.getLength(); i++)
 				{
 					assertEquals(constArr2[i], i + 1000);
 				}
 
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					arr[i] = i + 2000;
 				}
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					assertEquals(arr[i], i + 2000);
 				}
-				for (int i = 0; i < constArr2.getLength(); i++)
+				for (size_t i = 0; i < constArr2.getLength(); i++)
 				{
 					assertEquals(constArr2[i], i + 1000);
 				}
 
 				bbe::Array<int, 16> arr2 = std::move(arr);
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i], i + 2000);
 				}
 
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					arr2[i] = i + 3000;
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i], i + 3000);
 				}
@@ -74,36 +74,36 @@ namespace bbe
 				bbe::Array<int, 16> arr3;
 				arr3 = arr2;
 
-				for (int i = 0; i < arr3.getLength(); i++)
+				for (size_t i = 0; i < arr3.getLength(); i++)
 				{
 					assertEquals(arr3[i], i + 3000);
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i], i + 3000);
 				}
 
-				for (int i = 0; i < arr3.getLength(); i++)
+				for (size_t i = 0; i < arr3.getLength(); i++)
 				{
 					arr3[i] = i + 4000;
 				}
-				for (int i = 0; i < arr3.getLength(); i++)
+				for (size_t i = 0; i < arr3.getLength(); i++)
 				{
 					assertEquals(arr3[i], i + 4000);
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i], i + 3000);
 				}
 
 				arr2 = std::move(arr3);
 
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i], i + 4000);
 				}
 
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i], i + 4000);
 				}
@@ -114,7 +114,7 @@ namespace bbe
 				assertEquals(hash(arr4), hash(arr2));
 
 				bbe::Array<int, 4> arr5{ 1, 2, 3, 4 };
-				for (int i = 0; i < arr5.getLength(); i++)
+				for (size_t i = 0; i < arr5.getLength(); i++)
 				{
 					assertEquals(arr5[i], i + 1);
 				}
@@ -131,54 +131,54 @@ namespace bbe
 				assertUnequals(constArr.getRaw(), nullptr);
 				assertEquals(constArr.getLength(), 16);
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					arr[i] = Person("", "", i + 1000);
 				}
 				
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					assertEquals(arr[i].age, i + 1000);
 				}
 
 				const bbe::Array<Person, 16> constArr2 = arr;
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					assertEquals(arr[i].age, i + 1000);
 				}
-				for (int i = 0; i < constArr2.getLength(); i++)
+				for (size_t i = 0; i < constArr2.getLength(); i++)
 				{
 					assertEquals(constArr2[i].age, i + 1000);
 				}
 
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					arr[i] = Person("", "", i + 2000);
 				}
 
-				for (int i = 0; i < arr.getLength(); i++)
+				for (size_t i = 0; i < arr.getLength(); i++)
 				{
 					assertEquals(arr[i].age, i + 2000);
 				}
-				for (int i = 0; i < constArr2.getLength(); i++)
+				for (size_t i = 0; i < constArr2.getLength(); i++)
 				{
 					assertEquals(constArr2[i].age, i + 1000);
 				}
 
 				bbe::Array<Person, 16> arr2 = std::move(arr);
 
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i].age, i + 2000);
 				}
 
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					arr2[i] = Person("", "", i + 3000);
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i].age, i + 3000);
 				}
@@ -186,35 +186,35 @@ namespace bbe
 				bbe::Array<Person, 16> arr3;
 				arr3 = arr2;
 
-				for (int i = 0; i < arr3.getLength(); i++)
+				for (size_t i = 0; i < arr3.getLength(); i++)
 				{
 					assertEquals(arr3[i].age, i + 3000);
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i].age, i + 3000);
 				}
 
-				for (int i = 0; i < arr3.getLength(); i++)
+				for (size_t i = 0; i < arr3.getLength(); i++)
 				{
 					arr3[i] = Person("", "", i + 4000);
 				}
-				for (int i = 0; i < arr3.getLength(); i++)
+				for (size_t i = 0; i < arr3.getLength(); i++)
 				{
 					assertEquals(arr3[i].age, i + 4000);
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i].age, i + 3000);
 				}
 
 				arr2 = std::move(arr3);
 
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i].age, i + 4000);
 				}
-				for (int i = 0; i < arr2.getLength(); i++)
+				for (size_t i = 0; i < arr2.getLength(); i++)
 				{
 					assertEquals(arr2[i].age, i + 4000);
 				}
