+++ BrotBoxEngineTest/Tests/GeneralPurposeAllocatorTest.h
@@ -74,7 +74,7 @@ namespace bbe {
 					}
 				}
 
-				for (int i = 0; i < list.getLength(); i++)
+				for (size_t i = 0; i < list.getLength(); i++)
 				{
 					gpa.deallocate(list[i]);
 				}
